ciao-deploy: update documentation to use privileged container
As the container use `/usr/bin/losetup` to manipulate the
CNCI image, it's neccessary to use the container in
`--privileged` mode and map `/dev/`, otherwise the container
won't be able to perfomr `losetup` operation nor mount the loops.

this commit updates the documentation to instruct the user how
to call the container in privileged mode.

This commit fixes #16

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>