ciao-deploy: don't install roles from ansible galaxy
This is a cleanup commit; since the documentation ask the user
to map the github.com/01org/ciao repository directly into the
ciao-deploy container, it's not neccessary to install the roles
from ansible-galaxy anymore due to those roles being used by
ansible directly from the ciao repo.

The `openstack-common` bundle is removed from the
`swupd bundle-add` command because the `openstack-python-clients`
bundle already includes it implicit.

This commit fixes #14

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>