Create clr-sdk Docker Container.
Docker container for running Clear Linux developer tooling.

Allows a user to run mixer tool in a cross-platform way.

Built atop the clearlinux:latest image, and includes the
os-clr-on-clr bundle for devloper tools. A python script
sets up up a non-root user (with UID & GID mapped to the
mix directory mounted at runtime), with sudo privileges
only for mixer, swupd, and mock.

Note: The README mentions privileges needed to run mixer
build-image. This is sufficient for now, and narrowing
the --privileged to the specific CAP subset (SYS_ADMIN,
MKNOD, etc?) is left for a future task.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>