Add clearlinux/mixer-ci Docker Image
This patch introduces an image used for running the Mixer CI on Travis.

Previously, a very large part of the time Travis spent building and
validating Mixer PR's was spent 'swupd bundle-add'ing the packages
needed. The _mixer_ build was very short once this step was done.

This image removes the need to build and configure the image with each
Travis run; only the code needs to be copied over before the mixer
build and tests can run.

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>