Add clearlinux/clr-installer-ci Docker Image
This pull request introduces an image to be used for running
the Clear Linux OS Installer continuous integration on Travis.

Previously, a very large part of the test time in Travis was
spent building adding the required bundle to the standard
Clear Linux Docker image to properly setup the test environment.

Please ensure that auto-trigger is enabled so that this image
is automatically rebuilt with each new base clearlinux docker image.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>