Update README.md for Clear Linux Stacks launch
Removing parts of the description that do not add value and supplying a
legally-approved description. Making formatting changes to enhance the
appearance of the long description. This will ensure that when automated
builds occur on Docker Hub that the description remains correct and does
not revert to an older, non-approved version.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>