Hide the expected warning msg
When removing the duplicated file and directory via rm -d,
if it's a non-empty directory, it will prompt the warning
message that the directory could not be removed becuase it
is not empty. This is as expectation, as there are other
files under the same diretory and we need to keep them.
Hide these warning message to avoid confusion.

Signed-off-by: Liu, Jianjun <jianjun.liu@intel.com>