Add DBRS dockerfiles
This commit includes the dockerfiles and documentation for Database
Reference Stack v0.1.0, they will be located under stacks/dbrs
directory. This version includes Redis and Cassandra applications.

Each application includes the scripts and conf directores with the
files required for building the container.

For Cassandra there is a helm chart, and for Redis there is a yaml
configuration file for a redisfailover instance. These files are
included for an easier adoption as a kubernetes application.

Signed-off-by: Gabriel Briones <gabriel.briones.sayeg@intel.com>